chore(scm-multi-platform-detection): Removing legacy platform detector
