+++ src/sentry/integrations/api/endpoints/organization_repository_platforms.py
@@ -11,7 +11,6 @@ from sentry.api.base import cell_silo_endpoint
 from sentry.integrations.api.bases.organization_repository import OrganizationRepositoryEndpoint
 from sentry.integrations.github.client import GitHubApiClient
 from sentry.integrations.github.multi_platform_detection import detect_platforms_multi
-from sentry.integrations.github.platform_detection import detect_platforms
 from sentry.integrations.services.integration import integration_service
 from sentry.integrations.types import IntegrationProviderSlug
 from sentry.models.organization import Organization
@@ -19,10 +18,9 @@ from sentry.models.repository import Repository
 from sentry.shared_integrations.exceptions import ApiConflictError, ApiError
 
 
-def _capture_detection_exception(type: str, is_multi: bool, repo_id: int, repo_name: str) -> None:
+def _capture_detection_exception(type: str, repo_id: int, repo_name: str) -> None:
     with sentry_sdk.new_scope() as scope:
         scope.set_tag("scm_platform_detection", type)
-        scope.set_tag("is_multi", is_multi)
         scope.set_tag("repo_id", repo_id)
         scope.set_tag("repo_name", repo_name)
         sentry_sdk.capture_exception()
@@ -67,22 +65,14 @@ class OrganizationRepositoryPlatformsEndpoint(OrganizationRepositoryEndpoint):
 
         client = GitHubApiClient(integration=integration, org_integration_id=org_integration.id)
 
-        is_multi = features.has(
-            "organizations:integrations-github-multi-platform-detection",
-            organization,
-            actor=request.user,
-        )
         try:
-            if is_multi:
-                platforms = detect_platforms_multi(client, repo.name)["platforms"]
-            else:
-                platforms = detect_platforms(client=client, repo=repo.name)
+            platforms = detect_platforms_multi(client, repo.name)["platforms"]
         except ApiConflictError:
-            # Empty / unprocessable repo (e.g. empty git tree) — multi path only.
-            _capture_detection_exception("empty_repo", is_multi, repo.id, repo.name)
+            # Empty / unprocessable repo (e.g. empty git tree).
+            _capture_detection_exception("empty_repo", repo.id, repo.name)
             return Response({"platforms": []})
         except (ApiError, ValueError):
-            _capture_detection_exception("failed", is_multi, repo.id, repo.name)
+            _capture_detection_exception("failed", repo.id, repo.name)
             return Response({"detail": "Failed to detect platforms from GitHub."}, status=502)
 
         return Response({"platforms": platforms})
